Add feature to comment on portfolio using POST method
