+++ portfolio/src/main/webapp/style.css
@@ -11,6 +11,9 @@ p {
 .page-section {
   padding: 3rem 0;
 }
+.bglight {
+  background-color: #f5f5f5
+}
 .page-section h2.section-heading {
   font-size: 2.5rem;
   margin-top: 0;
